Renamed build.h to platformid.h
