+++ libraries/commonutils/src/buildenums.h
@@ -16,7 +16,7 @@ GNU General Public License for more details.
 #ifndef BUILDENUMS_H
 #define BUILDENUMS_H
 
-#include "PlatformDefs/build.h"
+#include "PlatformDefs/platformid.h"
 
 // This header defines the enumeration values that can be passed to Q_build*
 // functions and get current value through XASH_PLATFORM, XASH_ARCHITECTURE and
