+++ xash3d_engine/engine/client/avi/avi_stub.c
@@ -13,7 +13,7 @@ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.
 */
 
-#include "PlatformDefs/build.h"
+#include "PlatformDefs/platformid.h"
 #include "common.h"
 
 #if !XASH_WIN32
