+++ xash3d_engine/engine/common/host.c
@@ -13,7 +13,7 @@ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.
 */
 
-#include "PlatformDefs/build.h"
+#include "PlatformDefs/platformid.h"
 #ifdef XASH_SDL
 #include <SDL.h>
 #endif  // XASH_SDL
