+++ xash3d_engine/engine/common/launcher.c
@@ -15,7 +15,7 @@ GNU General Public License for more details.
 
 #ifdef SINGLE_BINARY
 
-#include "PlatformDefs/build.h"
+#include "PlatformDefs/platformid.h"
 #include "common.h"
 #include "PlatformLib/System.h"
