+++ xash3d_engine/game_launch/game.cpp
@@ -13,7 +13,7 @@ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.
 */
 
-#include "PlatformDefs/build.h"
+#include "PlatformDefs/platformid.h"
 
 #include <stdio.h>
 #include <string.h>
